Separated responsibilities and created new folders for specific components
