+++ webapp/src/components/PrivateRoute.js
@@ -1,11 +1,11 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
-import { useAuth } from './AuthContext';
+import useAuth from '../hooks/useAuth';
 
 const PrivateRoute = () => {
   const { auth } = useAuth();
   const location = useLocation();
 
-  return auth ? <Outlet /> : <Navigate to="/login" state={{ from: location }} replace />;
+  return auth?.username ? <Outlet /> : <Navigate to="/login" state={{ from: location }} replace />;
 };
 
 export default PrivateRoute;
