+++ webapp/src/hooks/useAuth.js
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import AuthContext from "../context/AuthContext";
+
+const useAuth = () => {
+    return useContext(AuthContext);
+}
+
+export default useAuth;
